+++ artifact_gen/lib/component/from_map.dart
@@ -45,6 +45,48 @@ class $ArtifactFromMapComponent implements $ArtifactBuilderOutput {
     buf.writeln(
       "  static ${builder.applyDefsF(clazz.name)} fromProperties(String j)=>fromMap(${builder.applyDefsF("ArtifactCodecUtil")}.g(j));",
     );
+
+    buf.write(
+      "  static ${builder.applyDefsF(clazz.name)} get newInstance=>${builder.applyDefsF(clazz.name)}(",
+    );
+
+    for (ParameterElement i in params) {
+      if (i.isRequired) {
+        if (i.type.nullabilitySuffix == NullabilitySuffix.question) {
+          buf.write("${i.name}: null,");
+        } else if (i.hasDefaultValue) {
+          String defaultCode =
+              i.defaultValueCode == null
+                  ? 'null'
+                  : builder.valD(i.defaultValueCode.toString(), i.type);
+          buf.write("${i.name}: $defaultCode,");
+        } else if (i.type.isDartCoreBool) {
+          buf.write("${i.name}: _F,");
+        } else if (i.type.isDartCoreInt || i.type.isDartCoreDouble) {
+          buf.write("${i.name}: 0,");
+        } else if (i.type.isDartCoreString) {
+          buf.write("${i.name}: '',");
+        } else if (i.type.isDartCoreIterable || i.type.isDartCoreList) {
+          buf.write("${i.name}: [],");
+        } else if (i.type.name == "DateTime") {
+          buf.write("${i.name}: DateTime.now(),");
+        } else if (i.type.isDartCoreEnum ||
+            (i.type is InterfaceType && $isEnum(i.type as InterfaceType))) {
+          buf.write(
+            "${i.name}: ${builder.applyDefsF(i.type.getDisplayString(withNullability: false))}.values.first,",
+          );
+        } else if (i.type.isDartCoreSet || i.type.isDartCoreMap) {
+          buf.write("${i.name}: {},");
+        } else {
+          buf.write(
+            "${i.name}: \$${(i.type.getDisplayString(withNullability: false))}.newInstance,",
+          );
+        }
+      }
+    }
+
+    buf.writeln(');');
+
     buf.write(
       '  static ${builder.applyDefsF(clazz.name)} fromMap(${builder.applyDefsF("Map<String, dynamic>")} r){',
     );
@@ -143,4 +185,7 @@ class $ArtifactFromMapComponent implements $ArtifactBuilderOutput {
     buf.writeln('}');
     return (importUris, buf);
   }
+
+  bool $isEnum(InterfaceType type) =>
+      type.element is EnumElement || type.element.kind == ElementKind.ENUM;
 }