+++ example/lib/example.dart
@@ -1,22 +1,42 @@
 import 'package:artifact/artifact.dart';
 import 'package:example/gen/artifacts.gen.dart';
 
-@Artifact(compression: false, reflection: false, generateSchema: false)
+@artifact
 class Test {
-  String name;
-  @rename("agex")
-  int? age;
+  final String aString;
+  final int aInt;
+  final AEnum anEnum;
+  final Sub aSub;
+  final List<Sub>? subList;
+  final Set<int> aSet;
+  final Map<int, String> aMap;
 
-  Test({this.name = "John Doe"});
+  const Test({
+    this.aString = 'default',
+    required this.aInt,
+    required this.anEnum,
+    required this.aSub,
+    required this.subList,
+    required this.aSet,
+    required this.aMap,
+  });
 }
 
-void main() {
-  Test inst = Test();
-  print(inst.toJson(pretty: true));
-  inst.name = "NewName";
-  inst.age = 3;
-  print(inst.toJson(pretty: true));
+@artifact
+class Sub {
+  final int value;
 
-  inst = $Test.fromJson(inst.toJson());
-  print(inst.toJson(pretty: true));
+  Sub({this.value = 42});
+}
+
+enum AEnum { first, second, third }
+
+void main() {
+  Map<String, dynamic> data = {"value": 41};
+
+  Object object = $constructArtifact<Sub>();
+
+  print(object.runtimeType);
+  print($isArtifact(object));
+  print((object as Sub).toYaml());
 }